requirements.txt
# Dependências do projeto Arbitron

# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Caching
cachetools==5.3.2

# Web framework (Flask)
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1

# Notifications
python-telegram-bot==20.7

# Data processing
pandas==2.1.4
numpy==1.25.2

# Monitoring and logging
structlog==23.2.0

